+++ packages/xchain-bitcoin/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './client-types'
 export * from './common'
-export * from './ledger'
+export * from './sochain-api-types'
+export * from './haskoin-api-types'