+++ packages/xchain-bitcoin/src/utils.ts
@@ -20,6 +20,7 @@ import { BTC_DECIMAL, MIN_TX_FEE } from './const'
 import * as haskoinApi from './haskoin-api'
 import * as sochain from './sochain-api'
 import { BroadcastTxParams, UTXO } from './types/common'
+import * as HaskoinApiTypes from './types/haskoin-api-types'
 import { AddressParams, BtcAddressUTXO, ScanUTXOParam } from './types/sochain-api-types'
 
 const TX_EMPTY_SIZE = 4 + 1 + 1 + 4 //10
@@ -104,21 +105,29 @@ export const btcNetwork = (network: Network): Bitcoin.Network => {
  * @param {Address} address
  * @returns {Balance[]} The balances of the given address.
  */
-export const getBalance = async (params: AddressParams, haskoinUrl: string): Promise<Balance[]> => {
+export const getBalance = async ({
+  params,
+  haskoinUrl,
+  confirmedOnly,
+}: {
+  params: AddressParams
+  haskoinUrl: string
+  confirmedOnly: boolean
+}): Promise<Balance[]> => {
   switch (params.network) {
     case Network.Mainnet:
     case Network.Stagenet:
       return [
         {
           asset: AssetBTC,
-          amount: await haskoinApi.getBalance({ haskoinUrl, address: params.address }),
+          amount: await haskoinApi.getBalance({ haskoinUrl, address: params.address, confirmedOnly }),
         },
       ]
     case Network.Testnet:
       return [
         {
           asset: AssetBTC,
-          amount: await sochain.getBalance(params),
+          amount: await sochain.getBalance({ ...params, confirmedOnly }),
         },
       ]
   }
@@ -140,6 +149,29 @@ export const validateAddress = (address: Address, network: Network): boolean => 
   }
 }
 
+// Stores list of txHex in memory to avoid requesting same data
+const txHexMap: Record<TxHash, string> = {}
+
+type GetTxHexParams = {
+  sochainUrl: string
+  txHash: TxHash
+  network: Network
+}
+
+/**
+ * Helper to use cached `txHex` values
+ */
+const getTxHexFromCache = async ({ txHash, sochainUrl, network }: GetTxHexParams): Promise<string> => {
+  // try to get it from cache
+  let txHex = txHexMap[txHash]
+  if (txHex) return txHex
+  // or get it from Sochain
+  txHex = (await sochain.getTx({ hash: txHash, sochainUrl, network })).tx_hex
+  // cache it
+  txHexMap[txHash] = txHex
+  return txHex
+}
+
 /**
  * Scan UTXOs from sochain.
  *
@@ -154,33 +186,23 @@ export const scanUTXOs = async ({
   network,
   address,
   confirmedOnly = true, // default: scan only confirmed UTXOs
-  fetchTxHex,
+  withTxHex = false,
 }: ScanUTXOParam): Promise<UTXO[]> => {
   switch (network) {
     case Network.Testnet: {
-      let utxos: BtcAddressUTXO[] = []
-
       const addressParam: AddressParams = {
         sochainUrl,
         network,
         address,
       }
 
-      if (confirmedOnly) {
-        utxos = await sochain.getConfirmedUnspentTxs(addressParam)
-      } else {
-        utxos = await sochain.getUnspentTxs(addressParam)
-      }
+      // Get UTXOs from Sochain
+      const utxos: BtcAddressUTXO[] = confirmedOnly
+        ? await sochain.getConfirmedUnspentTxs(addressParam)
+        : await sochain.getUnspentTxs(addressParam)
 
-      const results: UTXO[] = []
-
-      for (const utxo of utxos) {
-        let txHex
-        if (fetchTxHex) {
-          txHex = (await sochain.getTx({ hash: utxo.txid, sochainUrl, network })).tx_hex
-        }
-
-        results.push({
+      return await Promise.all(
+        utxos.map(async (utxo) => ({
           hash: utxo.txid,
           index: utxo.output_no,
           value: assetToBase(assetAmount(utxo.value, BTC_DECIMAL)).amount().toNumber(),
@@ -188,33 +210,28 @@ export const scanUTXOs = async ({
             value: assetToBase(assetAmount(utxo.value, BTC_DECIMAL)).amount().toNumber(),
             script: Buffer.from(utxo.script_hex, 'hex'),
           },
-          txHex,
-        })
-      }
-
-      return results
+          txHex: withTxHex ? await getTxHexFromCache({ txHash: utxo.txid, sochainUrl, network }) : undefined,
+        })),
+      )
     }
     case Network.Mainnet:
     case Network.Stagenet: {
-      let utxos: haskoinApi.UtxoData[] = []
+      // Get UTXOs from Haskoin
+      const utxos: HaskoinApiTypes.UtxoData[] = confirmedOnly
+        ? await haskoinApi.getConfirmedUnspentTxs({ address, haskoinUrl, sochainUrl, network })
+        : await haskoinApi.getUnspentTxs({ address, haskoinUrl })
 
-      if (confirmedOnly) {
-        utxos = await haskoinApi.getConfirmedUnspentTxs({ address, haskoinUrl })
-      } else {
-        utxos = await haskoinApi.getUnspentTxs({ address, haskoinUrl })
-      }
-
-      return utxos.map(
-        (utxo) =>
-          ({
-            hash: utxo.txid,
-            index: utxo.index,
+      return await Promise.all(
+        utxos.map(async (utxo) => ({
+          hash: utxo.txid,
+          index: utxo.index,
+          value: baseAmount(utxo.value, BTC_DECIMAL).amount().toNumber(),
+          witnessUtxo: {
             value: baseAmount(utxo.value, BTC_DECIMAL).amount().toNumber(),
-            witnessUtxo: {
-              value: baseAmount(utxo.value, BTC_DECIMAL).amount().toNumber(),
-              script: Buffer.from(utxo.pkscript, 'hex'),
-            },
-          } as UTXO),
+            script: Buffer.from(utxo.pkscript, 'hex'),
+          },
+          txHex: withTxHex ? await getTxHexFromCache({ txHash: utxo.txid, sochainUrl, network }) : undefined,
+        })),
       )
     }
   }
@@ -235,7 +252,7 @@ export const buildTx = async ({
   sochainUrl,
   haskoinUrl,
   spendPendingUTXO = false, // default: prevent spending uncomfirmed UTXOs
-  fetchTxHex = false,
+  withTxHex = false,
 }: TxParams & {
   feeRate: FeeRate
   sender: Address
@@ -243,11 +260,11 @@ export const buildTx = async ({
   sochainUrl: string
   haskoinUrl: string
   spendPendingUTXO?: boolean
-  fetchTxHex?: boolean
+  withTxHex?: boolean
 }): Promise<{ psbt: Bitcoin.Psbt; utxos: UTXO[]; inputs: UTXO[] }> => {
   // search only confirmed UTXOs if pending UTXO is not allowed
   const confirmedOnly = !spendPendingUTXO
-  const utxos = await scanUTXOs({ sochainUrl, haskoinUrl, network, address: sender, confirmedOnly, fetchTxHex })
+  const utxos = await scanUTXOs({ sochainUrl, haskoinUrl, network, address: sender, confirmedOnly, withTxHex })
 
   if (utxos.length === 0) throw new Error('No utxos to send')
   if (!validateAddress(recipient, network)) throw new Error('Invalid address')